Struct asn1_der::CopyingSource

pub struct CopyingSource<S: Source, U: Sink> { /* fields omitted */ }

A source that also copies each read element to the sink

Warning: if a call to write fails, this source will return an error and the element that has been read from the underlying source will be lost

Implementations

impl<S: Source, U: Sink> CopyingSource<S, U>

pub fn copy_next(&mut self) -> Result<(), Asn1DerError>

Copies the next element

pub fn copy_n(&mut self, n: usize) -> Result<(), Asn1DerError>

Copies the next n elements