Struct asn1_der::CountingSource

pub struct CountingSource<'a, S: Source> { /* private fields */ }

A source that counts the amount of elements read

Warning: if a call to read would cause ctr to exceed usize::max_value(), this source will return an error and the element that has been read from the underlying source will be lost