Trait asn1_der::IntoDerObject

pub trait IntoDerObject where
    Self: Sized, {
    fn into_der_object(self) -> DerObject;

    fn serialized_len(&self) -> usize;

    fn serialize<'a>(
        self, 
        buf: impl Iterator<Item = &'a mut u8>
    ) -> Result<usize, Asn1DerError> { ... }
}

A trait for converting native elements into a DER object

Required methods

fn into_der_object(self) -> DerObject

Converts self into a DER object

fn serialized_len(&self) -> usize

Efficiently computes the length of the DER-serialized representation of self

Loading content...
Provided methods

fn serialize<'a>(
self,
buf: impl Iterator<Item = &'a mut u8>
) -> Result<usize, Asn1DerError>

DER-serializes self into buf and returns the amount of bytes written

Loading content...