Struct asn1_der::typed::Sequence

pub struct Sequence<'a> { /* private fields */ }

An ASN.1-DER sequence view

Implementations

impl<'a> Sequence<'a>

pub fn new<S: Sink + Into<&'a [u8]>, T: DerEncodable>(
objs: &[T],
sink: S
) -> Result<Self, Asn1DerError>

Writes a new sequence object with objs as subobjects into sink and returns a type view over it

pub fn len(&self) -> usize

The amount of subelements in the sequence

Note: since there is no underlying index, the amount of subelements has to be recomputed every time. If you need the length more than once, consider caching it.

pub fn get(&self, n: usize) -> Result<DerObject<'a>, Asn1DerError>

Gets the nth subobject

Note: since there is no underlying index, the position of each subelement has to be recomputed every time. If you need the subobjects more than once, consider caching them.

pub fn get_as<T: DerDecodable<'a>>(&self, n: usize) -> Result<T, Asn1DerError>

Gets the nth subobject as T

Note: since there is no underlying index, the position of each subelement has to be recomputed every time. If you need the subobjects more than once, consider caching them.

pub fn write<S: Sink, T: DerEncodable>(
objs: &[T],
sink: &mut S
) -> Result<(), Asn1DerError>

Writes a sequence consisting of objs as DER-object to sink